9Xtreme fixes + USB joystick. Language additions

+++ radio/ersky9x/src/menus.h
@@ -164,7 +164,7 @@ extern int16_t scaleAnalog( int16_t v, uint8_t channel ) ;
 
 const char *get_curve_string() ;
 
-extern int16_t calc_scaler( uint8_t index, uint8_t *unit, uint8_t *num_decimals) ;
+extern int16_t calc_scaler( uint8_t index, uint16_t *unit, uint8_t *num_decimals) ;
 
 extern uint8_t CalcScaleNest ;
